less 60: @Query(value = "", nativeQuery = true)  <- nativeQuery @Query("select p from Citizen p where p.address.city = :cityName")  <- JPQL query @Transactional <- to update or any change methods into ServiceImpl @Transactional(readOnly = true) <- for read only methods method getCitiesPopulation with JPQL

+++ src/main/java/telran/java52/person/dto/CityPopulationDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CityPopulationDto {
 	String city;
-	Integer population;
+	Long population;
 }
